% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp_ppmn.R
\name{disp_ppmn}
\alias{disp_ppmn}
\title{Displacement test for PPMN}
\usage{
disp_ppmn(
  object,
  new_data,
  nsim = 3000,
  level = 0.9,
  gd_display = F,
  seed = 123
)
}
\arguments{
\item{object}{A foundational or updated PPMN}

\item{new_data}{New dataset with all root nodes and least one child node present}

\item{nsim}{Number of simulations needed for updating}

\item{level}{Credibility level (default level = 0.9)}

\item{gd_display}{Display the Generalized Displacement (GD) in the name}

\item{seed}{Seed value 123.}
}
\description{
Displacement test for PPMN
}
