% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmn_lambda_level.R
\name{.ppmn_lambda_level}
\alias{.ppmn_lambda_level}
\title{Helper function finding fraction of loss}
\usage{
.ppmn_lambda_level(loss, best_frac, max_lambda)
}
\arguments{
\item{loss}{A vector of losses.}

\item{max_lambda}{Maximum lambda.}

\item{level}{The level of loss.}
}
\description{
Helper function finding fraction of loss
}
\keyword{internal}
