% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmn_oob_lambda_pred.R
\name{.ppmn_oob_lambda_pred}
\alias{.ppmn_oob_lambda_pred}
\title{Helper OOB optimizer}
\usage{
.ppmn_oob_lambda_pred(obs_mat, loss, max_lambda, seed, seq_len, n_boot = 1000)
}
\arguments{
\item{obs_mat}{Observed matrix.}

\item{loss}{Loss matrix.}

\item{max_lambda}{Maximum lambda.}

\item{seed}{seed.}

\item{seq_len}{Length of lambda sequence.}

\item{n_boot}{number of bootstraps.}
}
\description{
Helper OOB optimizer
}
\keyword{internal}
