% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmn_resid_bboot.R
\name{.ppmn_resid_bboot}
\alias{.ppmn_resid_bboot}
\title{Helper function for Bayesian bootstrap}
\usage{
.ppmn_resid_bboot(resid1, resid0, n_grid = 3000, n_boot = 1000)
}
\arguments{
\item{resid1}{Residuals for particle with best fit.}

\item{resid0}{Residuals for the prior.}

\item{n_grid}{Size of the grid.}

\item{n_boot}{Number of bootstraps.}
}
\description{
Helper function for Bayesian bootstrap
}
\keyword{internal}
