% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmn_resid_mc_kl.R
\name{.ppmn_resid_mc_kl}
\alias{.ppmn_resid_mc_kl}
\title{Helper MC approx. for mv KL-divergence}
\usage{
.ppmn_resid_mc_kl(resid1, resid0, loss, n_boot = 1000, n_mc = 5000)
}
\arguments{
\item{resid1}{Residuals.}

\item{resid0}{Model prior residuals}

\item{loss}{Loss.}

\item{n_boot}{Number of bayesian bootstraps.}

\item{n_mc}{Number of mc.}
}
\description{
Helper MC approx. for mv KL-divergence
}
\keyword{internal}
