% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmn_stable_weights.R
\name{.ppmn_stable_weights}
\alias{.ppmn_stable_weights}
\title{Helper function to derive GBU weights}
\usage{
.ppmn_stable_weights(loss, lambda, prior = NULL)
}
\arguments{
\item{loss}{The loss.}

\item{lambda}{Lambda.}

\item{prior}{Prior.}
}
\description{
Helper function to derive GBU weights
}
\keyword{internal}
