% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmn_wquantile.R
\name{.ppmn_wquantile}
\alias{.ppmn_wquantile}
\title{Helper function for to calculated weighted quantiles.}
\usage{
.ppmn_wquantile(x, w, probs)
}
\arguments{
\item{x}{Values (particles/simulations).}

\item{w}{GBU weights.}

\item{probs}{Probabilities}
}
\description{
Helper function for to calculated weighted quantiles.
}
