% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example4}
\alias{example4}
\title{Example meta-analytic data for a single PPMN edge}
\format{
A data frame containing meta-analytic information, including:
\describe{
\item{Source}{Identifier for the source or fitted model.}
\item{Authors}{Authors associated with the source.}
\item{Comment}{Additional information about the source or model.}
\item{function}{Functional form used for the edge-function.}
\item{transformation}{Transformation applied to the predictor variable.}
\item{edge}{Directed relationship represented as a model formula.}
\item{parameter}{Model parameter associated with the estimate.}
\item{estimate}{Estimated value of the model parameter.}
\item{error}{Standard error associated with the parameter estimate.}
}
}
\usage{
example4
}
\description{
Example meta-analytic dataset containing parameter estimates and
associated uncertainty for constructing a simple Posterior Predictive
Meta-analytic Network (PPMN).
}
\details{
The dataset contains estimates from fitted models describing the
relationship between total phosphorus (\code{TP}) and chlorophyll-a (\code{Chl}).
It is used in the HMiE examples to demonstrate the construction of a
foundational PPMN.
}
\examples{
data(example4)
head(example4)

}
\keyword{datasets}
