% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example5}
\alias{example5}
\title{Example meta-analytic data for a multi-edge PPMN}
\format{
A data frame containing meta-analytic parameter estimates,
standard errors, edge definitions, functional forms, and associated
information used by \code{\link[=build_ppmn]{build_ppmn()}}.
}
\usage{
example5
}
\description{
Example meta-analytic dataset used to demonstrate construction of a
Posterior Predictive Meta-analytic Network (PPMN) containing multiple
ecological relationships.
}
\details{
The example includes relationships involving total nitrogen (\code{TN}),
total phosphorus (\code{TP}), chlorophyll-a (\code{Chl}), and phytoplankton
taxa richness (\code{Taxa}).
}
\examples{
data(example5)
head(example5)

}
\keyword{datasets}
