% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example6}
\alias{example6}
\title{Example meta-analytic data for a larger PPMN}
\format{
A data frame containing meta-analytic parameter estimates,
standard errors, edge definitions, functional forms, transformations,
and associated information used by \code{\link[=build_ppmn]{build_ppmn()}}.
}
\usage{
example6
}
\description{
Example meta-analytic dataset used to construct a larger ecological
Posterior Predictive Meta-analytic Network (PPMN).
}
\details{
The corresponding example network contains several ecological variables
and edge-functions and is used to demonstrate construction and subsequent
updating of a PPMN.
}
\examples{
data(example6)
head(example6)

}
\keyword{datasets}
