% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_ppmn.R
\name{marginal_ppmn}
\alias{marginal_ppmn}
\title{Plots the marginal expectations of each edge function with credibility intervals}
\usage{
marginal_ppmn(
  object,
  child,
  data,
  level = 0.9,
  nsim = 1000,
  n_grid = 100,
  smooth_f = 0.5,
  ylim = NULL,
  ylab = NULL,
  lwd_exp = 0.8,
  lwd_col = "dodgerblue4",
  size_title = 8,
  size_text = 8,
  int_col = "dodgerblue3",
  int_alpha = 0.3,
  pt_size = 2,
  pt_alpha = 0.25,
  seed = 123
)
}
\arguments{
\item{object}{A foundational or updated PPMN which will be used for predictions.}

\item{child}{The dependent (child) variable to display.}

\item{data}{Data that contains x and y coordinates (points) to plot.}

\item{level}{Credibility level (default level = 0.9).}

\item{nsim}{Number of simulations to construct the intervals.}

\item{n_grid}{Grid size of the added variable plot.}

\item{smooth_f}{A value for the smoother span only used to display specific non-linear models.}

\item{ylim}{Limit of the y-axis.}

\item{ylab}{Title of the y-axis.}

\item{lwd_exp}{Line width of the expected value.}

\item{lwd_col}{Colour of the expected value.}

\item{size_title}{Title size of the y-axis label.}

\item{size_text}{Text size of the title.}

\item{int_col}{Colour of the credibility intervals around the expected value.}

\item{int_alpha}{Transparency of the credibility intervals around the expected value.}

\item{pt_size}{Size of the points in the plot.}

\item{pt_alpha}{Transparency of the points in the plot.}

\item{seed}{Seed value 123.}
}
\description{
This is essentially an added variable plot. It displays the marginal relation of the
dependent variable and all independent variables for each edge-function belonging to
the selected child. If multiple edge-functions predict the same child, a separate
marginal plot is returned for every edge-function and parent combination.
}
