% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ppmn.R
\name{print.ppmn}
\alias{print.ppmn}
\title{Helper function to print summary table}
\usage{
\method{print}{ppmn}(x, ...)
}
\arguments{
\item{x}{An object to summarize.}
}
\description{
Helper function to print summary table
}
\keyword{internal}
