% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ppmn_predict.R
\name{print.ppmn_prediction}
\alias{print.ppmn_prediction}
\title{Helper function for predictions}
\usage{
\method{print}{ppmn_prediction}(x, ...)
}
\arguments{
\item{x}{An object to summarize.}
}
\description{
Helper function for predictions
}
