% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ppmn_root_estimation.R
\name{print.ppmn_root_estimation}
\alias{print.ppmn_root_estimation}
\title{Helper function for root estimation}
\usage{
\method{print}{ppmn_root_estimation}(x, ...)
}
\arguments{
\item{x}{An object to summarize.}
}
\description{
Helper function for root estimation
}
\keyword{internal}
