% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_data}
\alias{train_data}
\title{Example training data for updating a PPMN}
\format{
A data frame in which rows represent observations and columns
represent ecological variables corresponding to vertices in the example
PPMN.
}
\usage{
train_data
}
\description{
Multivariate ecological data used to demonstrate sequential updating
of a foundational Posterior Predictive Meta-analytic Network (PPMN).
}
\details{
The dataset provides observations for multiple vertices of the example
network and is used by \code{\link[=update_ppmn]{update_ppmn()}} to update the foundational PPMN
using multivariate information.
}
\examples{
data(train_data)
head(train_data)

}
\keyword{datasets}
