% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{val_data}
\alias{val_data}
\title{Example validation data for a PPMN}
\format{
A data frame in which rows represent observations and columns
represent ecological variables corresponding to vertices in the example
PPMN.
}
\usage{
val_data
}
\description{
Multivariate ecological data used to demonstrate predictive assessment
of an updated Posterior Predictive Meta-analytic Network (PPMN).
}
\details{
The dataset is intended for evaluating predictions after the example
network has been updated using \code{train_data}.
}
\examples{
data(val_data)
head(val_data)

}
\keyword{datasets}
