% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmn_detect_type.R
\name{.ppmn_detect_type}
\alias{.ppmn_detect_type}
\title{Helper function to detect dependent variable types}
\usage{
.ppmn_detect_type(y_raw)
}
\arguments{
\item{y_raw}{Raw observations.}
}
\description{
Helper function to detect dependent variable types
}
