% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmn_kl_calibrate.R
\name{.ppmn_kl_calibrate}
\alias{.ppmn_kl_calibrate}
\title{Grid approximation for Kullenback-Leibler divergence}
\usage{
.ppmn_kl_calibrate(r_data, r_prior, n_boot = 1000, n_grid, n_samp)
}
\arguments{
\item{r_data}{Residuals of the data.}

\item{r_prior}{Residuals of the prior.}

\item{n_boot}{Number of bootstraps.}

\item{n_grid}{Size of the grid.}
}
\description{
Grid approximation for Kullenback-Leibler divergence
}
\keyword{internal}
