% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmn_kl_optim.R
\name{.ppmn_kl_optim}
\alias{.ppmn_kl_optim}
\title{Helper function to Kullenback-Leibler divergence optimizer}
\usage{
.ppmn_kl_optim(lambda, loss, kl_target, prior = NULL)
}
\arguments{
\item{lambda}{Lambda}

\item{loss}{Loss}

\item{kl_target}{KL divergence target}

\item{prior}{Prior}
}
\description{
Helper function to Kullenback-Leibler divergence optimizer
}
\keyword{internal}
