% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmn_lambda_kl.R
\name{.ppmn_lambda_kl}
\alias{.ppmn_lambda_kl}
\title{Helper function for Kullenback-Leiber optimizer for lambda}
\usage{
.ppmn_lambda_kl(loss, kl_target, prior = NULL, upper = max_lambda)
}
\arguments{
\item{loss}{The loss.}

\item{kl_target}{KL-divergence target value.}

\item{prior}{Prior.}

\item{upper}{Upper boundary for lambda.}
}
\description{
Helper function for Kullenback-Leiber optimizer for lambda
}
\keyword{internal}
