% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmn_logistic_loss.R
\name{.ppmn_logistic_loss}
\alias{.ppmn_logistic_loss}
\title{Helper function for binary/prop log loss}
\usage{
.ppmn_logistic_loss(x_raw, x_hat)
}
\arguments{
\item{x_raw}{Raw observations.}

\item{x_hat}{Predictions on these observations.}
}
\description{
Helper function for binary/prop log loss
}
\keyword{internal}
