% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmn_loss_resid.R
\name{.ppmn_loss_resid}
\alias{.ppmn_loss_resid}
\title{Helper function to calculate loss and residuals}
\usage{
.ppmn_loss_resid(obs_mat, preds, obs_nodes, nsim, mad_constant = 1.4826)
}
\arguments{
\item{obs_mat}{Observed values matrix.}

\item{preds}{Predictions on this observed matrix.}

\item{obs_nodes}{Observed nodes of the matrix.}

\item{nsim}{Number of simulations.}

\item{mad_constant}{Constant.}
}
\description{
Helper function to calculate loss and residuals
}
\keyword{internal}
