% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmn_predict_edge.R
\name{.ppmn_predict_edge}
\alias{.ppmn_predict_edge}
\title{Helper function that defines and predicts with different functions}
\usage{
.ppmn_predict_edge(func, beta, trans, x)
}
\arguments{
\item{func}{Name of the function.}

\item{beta}{Parameters of the functions.}

\item{trans}{Transformation type applied over all independent variables.}

\item{x}{matrix of independent variables.}
}
\description{
Helper function that defines and predicts with different functions
}
\keyword{internal}
