% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmn_resid_diagnostic.R
\name{.ppmn_resid_diagnostic}
\alias{.ppmn_resid_diagnostic}
\title{Helper function to residual diagnostics}
\usage{
.ppmn_resid_diagnostic(object, data, mad_constant = 1.4826)
}
\arguments{
\item{object}{An updated PPMN.}

\item{data}{The data used to update the PPMN.}

\item{mad_constant}{Constant.}
}
\description{
Helper function to residual diagnostics
}
\keyword{internal}
