% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmn_tab_fun.R
\name{.ppmn_tab_fun}
\alias{.ppmn_tab_fun}
\title{Helper function to create parameter table}
\usage{
.ppmn_tab_fun(param_est, e_fun)
}
\arguments{
\item{param_est}{Parameter estimate}

\item{e_fun}{Edge function}
}
\description{
Helper function to create parameter table
}
\keyword{internal}
