% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmn_transform_loss.R
\name{.ppmn_transform_loss}
\alias{.ppmn_transform_loss}
\title{Helper transformation function}
\usage{
.ppmn_transform_loss(x_raw, x_hat, type, MAD = NULL)
}
\arguments{
\item{x_raw}{Raw observations.}

\item{x_hat}{Predictions on x_raw.}

\item{type}{Type of distribution for x_raw.}

\item{MAD}{Median absolute deviation.}
}
\description{
Helper transformation function
}
\keyword{internal}
