% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ppmn_pred_fit.R
\name{print.ppmn_pred_fit}
\alias{print.ppmn_pred_fit}
\title{Helper function for predictive fit}
\usage{
\method{print}{ppmn_pred_fit}(x, ...)
}
\arguments{
\item{x}{An object to summarize.}
}
\description{
Helper function for predictive fit
}
\keyword{internal}
