% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ppmn.R
\name{update_ppmn}
\alias{update_ppmn}
\title{Update PPMN function}
\usage{
update_ppmn(
  object,
  new_data,
  nsim = 3000,
  level = 0.9,
  method = "kernel",
  up_strength = 0.95,
  max_lambda = 1000,
  diagnostics = T,
  covar = T,
  seed = 123
)
}
\arguments{
\item{object}{A foundational or updated PPMN.}

\item{new_data}{New data set with all root vertices and least one child vertex present.}

\item{nsim}{Number of simulations used for updating.}

\item{level}{Generalized credibility level (default level = 0.9).}

\item{method}{The method being used to set the updating strength of the model. The default
method is 'kernel', The other method 'manual' uses the Kullenback-Leibler divergence applied
over the number of simulations to express the concentration strength relative to no change relative to the prior.
Here the default of the update strength is 95\%.}

\item{up_strength}{The fraction of information that is allowed to concentrate in the generalized posterior
when using the method 'manual'.}

\item{max_lambda}{Maximum lambda that can be set for Generalized Bayesian Updating (default max_lambda = 1000).}

\item{diagnostics}{If residual diagnostics should be returned (default diagnostics = T).}

\item{covar}{Covariance matrix that is calculated between model parameters (default covar = T).}

\item{seed}{Seed value 123.}
}
\description{
The updating function uses loss-based Generalized Bayesian Updating (GBU) to update the prior to a generalized
posterior (often called GIBBS posterior). In simplicity it simulates parameters (particles) from prior model
These particles are used in a function to predict. Each prediction the mismatch with the data (loss) is calculated.
To updated the prior these losses each particles is then weighted inversely relative to the losses.
}
